requirements.txt
streamlit
langchain
requests
google-cloud-language
textblob
groq
fastapi
uvicorn
python-dotenv
langchain-community